feat(temporal): adjust temporal workflows for realtime cohort calculation
