--- posthog/templates/admin/behavioral_cohort_analysis.html
@@ -1,189 +1,0 @@
-{% extends "admin/base_site.html" %}
-{% load i18n admin_urls static %}
-
-{% block extrastyle %}
-{{ super }}
-<style>
-    .form-row {
-        margin-bottom: 15px;
-    }
-    .help {
-        font-size: 11px;
-        color: #666;
-    }
-    .schedule-fields {
-        background: #f8f8f8;
-        padding: 15px;
-        border-radius: 5px;
-        margin-top: 20px;
-    }
-    .submit-row {
-        margin-top: 30px;
-    }
-    .alert {
-        padding: 10px;
-        margin: 20px 0;
-        border-radius: 5px;
-    }
-    .alert-info {
-        background: #e3f2fd;
-        border: 1px solid #90caf9;
-        color: #1565c0;
-    }
-</style>
-{% endblock %}
-
-{% block content %}
-
-<div class="alert alert-info">
-    <strong>Note:</strong> This tool triggers the Temporal workflow for analyzing behavioral cohorts. 
-    The analysis runs asynchronously in the background. Check the Temporal UI for progress and results.
-</div>
-
-{% if messages %}
-    {% for message in messages %}
-        <div class="alert alert-{{ message.tags }}">
-            {{ message }}
-        </div>
-    {% endfor %}
-{% endif %}
-
-<form method="post" id="behavioral-cohort-form">
-    {% csrf_token %}
-    
-    <fieldset class="module aligned">
-        <h2>Basic Configuration</h2>
-        
-        <div class="form-row">
-            <label for="{{ form.min_matches.id_for_label }}" class="required">
-                {{ form.min_matches.label }}:
-            </label>
-            {{ form.min_matches }}
-            {% if form.min_matches.help_text %}
-                <p class="help">{{ form.min_matches.help_text }}</p>
-            {% endif %}
-            {{ form.min_matches.errors }}
-        </div>
-        
-        <div class="form-row">
-            <label for="{{ form.days.id_for_label }}" class="required">
-                {{ form.days.label }}:
-            </label>
-            {{ form.days }}
-            {% if form.days.help_text %}
-                <p class="help">{{ form.days.help_text }}</p>
-            {% endif %}
-            {{ form.days.errors }}
-        </div>
-        
-        <div class="form-row">
-            <label for="{{ form.parallelism.id_for_label }}" class="required">
-                {{ form.parallelism.label }}:
-            </label>
-            {{ form.parallelism }}
-            {% if form.parallelism.help_text %}
-                <p class="help">{{ form.parallelism.help_text }}</p>
-            {% endif %}
-            {{ form.parallelism.errors }}
-        </div>
-    </fieldset>
-    
-    <fieldset class="module aligned">
-        <h2>Optional Filters</h2>
-        
-        <div class="form-row">
-            <label for="{{ form.team_id.id_for_label }}">
-                {{ form.team_id.label }}:
-            </label>
-            {{ form.team_id }}
-            {% if form.team_id.help_text %}
-                <p class="help">{{ form.team_id.help_text }}</p>
-            {% endif %}
-            {{ form.team_id.errors }}
-        </div>
-        
-        <div class="form-row">
-            <label for="{{ form.cohort_id.id_for_label }}">
-                {{ form.cohort_id.label }}:
-            </label>
-            {{ form.cohort_id }}
-            {% if form.cohort_id.help_text %}
-                <p class="help">{{ form.cohort_id.help_text }}</p>
-            {% endif %}
-            {{ form.cohort_id.errors }}
-        </div>
-        
-        <div class="form-row">
-            <label for="{{ form.condition.id_for_label }}">
-                {{ form.condition.label }}:
-            </label>
-            {{ form.condition }}
-            {% if form.condition.help_text %}
-                <p class="help">{{ form.condition.help_text }}</p>
-            {% endif %}
-            {{ form.condition.errors }}
-        </div>
-    </fieldset>
-    
-    <fieldset class="module aligned">
-        <h2>Scheduling Options</h2>
-        
-        <div class="form-row">
-            <label for="{{ form.schedule.id_for_label }}">
-                {{ form.schedule.label }}:
-            </label>
-            {{ form.schedule }}
-            {% if form.schedule.help_text %}
-                <p class="help">{{ form.schedule.help_text }}</p>
-            {% endif %}
-            {{ form.schedule.errors }}
-        </div>
-        
-        <div class="schedule-fields" id="schedule-fields" style="display: none;">
-            <div class="form-row">
-                <label for="{{ form.duration.id_for_label }}">
-                    {{ form.duration.label }}:
-                </label>
-                {{ form.duration }}
-                {% if form.duration.help_text %}
-                    <p class="help">{{ form.duration.help_text }}</p>
-                {% endif %}
-                {{ form.duration.errors }}
-            </div>
-            
-            <div class="form-row">
-                <label for="{{ form.interval.id_for_label }}">
-                    {{ form.interval.label }}:
-                </label>
-                {{ form.interval }}
-                {% if form.interval.help_text %}
-                    <p class="help">{{ form.interval.help_text }}</p>
-                {% endif %}
-                {{ form.interval.errors }}
-            </div>
-        </div>
-    </fieldset>
-    
-    <div class="submit-row">
-        <input type="submit" value="Run Analysis" class="default" />
-    </div>
-</form>
-
-<script>
-document.addEventListener('DOMContentLoaded', function() {
-    const scheduleCheckbox = document.querySelector('input[name="schedule"]');
-    const scheduleFields = document.getElementById('schedule-fields');
-    
-    function toggleScheduleFields() {
-        if (scheduleCheckbox.checked) {
-            scheduleFields.style.display = 'block';
-        } else {
-            scheduleFields.style.display = 'none';
-        }
-    }
-    
-    scheduleCheckbox.addEventListener('change', toggleScheduleFields);
-    toggleScheduleFields(); // Initial state
-});
-</script>
-{% endblock %}
