+++ posthog/templates/admin/realtime_cohort_calculation.html
@@ -29,9 +29,8 @@
 
 {% block content %}
 <div class="alert alert-info">
-    <strong>Note:</strong> This tool triggers the Temporal workflow for analyzing actions. 
-    Each action gets its own lightweight workflow for better reliability and observability.
-    The analysis runs asynchronously in the background and processes actions with bytecode.
+    <strong>Note:</strong> This tool triggers the Temporal workflow for realtime cohort calculation. 
+    The calculation runs asynchronously in the background and processes all actions with bytecode.
     Check the Temporal UI for progress and results.
 </div>
 
@@ -84,44 +83,33 @@
     </fieldset>
     
     <fieldset class="module aligned">
-        <h2>Batch Scheduling (New)</h2>
+        <h2>Jittered Scheduling</h2>
         
         <div class="form-row">
-            <label for="{{ form.batch_size.id_for_label }}" class="required">
-                {{ form.batch_size.label }}:
+            <label for="{{ form.workflows_per_batch.id_for_label }}" class="required">
+                {{ form.workflows_per_batch.label }}:
             </label>
-            {{ form.batch_size }}
-            {% if form.batch_size.help_text %}
-                <p class="help">{{ form.batch_size.help_text }}</p>
+            {{ form.workflows_per_batch }}
+            {% if form.workflows_per_batch.help_text %}
+                <p class="help">{{ form.workflows_per_batch.help_text }}</p>
             {% endif %}
-            {{ form.batch_size.errors }}
+            {{ form.workflows_per_batch.errors }}
         </div>
         
         <div class="form-row">
-            <label for="{{ form.batch_delay.id_for_label }}" class="required">
-                {{ form.batch_delay.label }}:
+            <label for="{{ form.batch_delay_minutes.id_for_label }}" class="required">
+                {{ form.batch_delay_minutes.label }}:
             </label>
-            {{ form.batch_delay }}
-            {% if form.batch_delay.help_text %}
-                <p class="help">{{ form.batch_delay.help_text }}</p>
+            {{ form.batch_delay_minutes }}
+            {% if form.batch_delay_minutes.help_text %}
+                <p class="help">{{ form.batch_delay_minutes.help_text }}</p>
             {% endif %}
-            {{ form.batch_delay.errors }}
-        </div>
-        
-        <div class="form-row">
-            <label for="{{ form.max_actions.id_for_label }}" class="required">
-                {{ form.max_actions.label }}:
-            </label>
-            {{ form.max_actions }}
-            {% if form.max_actions.help_text %}
-                <p class="help">{{ form.max_actions.help_text }}</p>
-            {% endif %}
-            {{ form.max_actions.errors }}
+            {{ form.batch_delay_minutes.errors }}
         </div>
     </fieldset>
     
     <div class="submit-row">
-        <input type="submit" value="Run Analysis" class="default" />
+        <input type="submit" value="Run Calculation" class="default" />
     </div>
 </form>
 {% endblock %}
