+++ posthog/templates/custom_tools/app.html
@@ -5,15 +5,9 @@
     </caption>
     <tr class="model-actionsanalysis">
       <th scope="row">
-        <a href="{% url 'actions-analysis' %}">Actions Analysis</a>
+        <a href="{% url 'realtime-cohorts-calculation' %}">Realtime Cohorts Calculation</a>
       </th>
-      <td>Run actions analysis workflow</td>
-    </tr>
-    <tr class="model-behavioralcohortanalysis">
-      <th scope="row">
-        <a href="{% url 'behavioral-cohort-analysis' %}">Behavioral Cohort Analysis</a>
-      </th>
-      <td>Run behavioral cohort analysis workflow</td>
+      <td>Run realtime cohorts calculation workflow</td>
     </tr>
   </table>
 </div>
